+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `PropertyContainer` now supports access via `[]` (Index)
 - Properties now provide convenience functions for type casting
 - Load most attributes on text objects (except for halign/valign)
+- Support for object templates
 
 ### Changed
 - ImageStorage::SpriteSheet now holds the image data as an Rc instead of a Box.